Create separate files for protocols
